doc: fix missing references to README.md
Cross-repo references using an URL that only has the folder name cause
links in the github.io site to go to the github.com site.  Most of these
uses want to link to the documentation (README.md) so add that where
appropriate so links in the github.io site stay there.

Also use an URL when relative links want to take you to the github.com
repo and add the README.md when the document is the link target.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>