[ROCm] Fix the hf-token setting for tgi and tei in chatqna
This PR is to correct the env variable name in chatqna example on ROCm platform passing to the docker container of tgi and tei. For tgi, either `HF_TOKEN` and `HUGGING_FACE_HUB_TOKEN` could be parsed in tgi while `HF_API_TOKEN` can be parsed in tei.

Fix the CI issue for ROCm in https://github.com/opea-project/GenAIExamples/pull/1430

Signed-off-by: Wang, Kai Lawrence <kai.lawrence.wang@intel.com>